copilot-agent: rename dir as copilot-agent
