+++ pkg/copilot-agent/nm/nm.go
@@ -31,7 +31,7 @@ import (
 	"github.com/koordinator-sh/koordinator/pkg/koordlet/util/system"
 	"k8s.io/klog/v2"
 
-	"github.com/koordinator-sh/goyarn/pkg/copilot/utils"
+	"github.com/koordinator-sh/goyarn/pkg/copilot-agent/utils"
 )
 
 const (
