+++ pkg/copilot-agent/nm/nm_pod_discover.go
