+++ pkg/copilot-agent/nm/types.go
