+++ pkg/copilot-agent/server/server.go
@@ -30,7 +30,7 @@ import (
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/klog/v2"
 
-	"github.com/koordinator-sh/goyarn/pkg/copilot/nm"
+	"github.com/koordinator-sh/goyarn/pkg/copilot-agent/nm"
 )
 
 type YarnCopilotServer struct {
